Implemented map of SSIM (window by window)
Now, this script read the two images and compare them window by window calculating the SSIM index of each window and storing them in a matrix.

The average of the values of this matrix is the MSSIM.

Furthermore, this matrix is mapped to the interval [0 , 255] and showed as an image.